ocr/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
Pillow>=10.0.0
pytesseract>=0.3.10
nltk>=3.8.1
werkzeug>=2.3.0 